backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
rapidfuzz==3.9.6
pytest==8.3.2
